Implemented alpha version of prediction system. When the ball is moving slowly the position is predicted acurately within a 5 point margin, I wasn't able to test it with our robots. (No key)
Things that need to be done:
testing of outputs, I made sure it shouldn't throw any errors, but predictions for bounces haven't been properly tested.
The values for the pitch boundaries, top, bottom, etc.  are hard coded, I didn't find where there are in the system, so I use hard coded values (bad) that are probably wrong (very bad).

The changes were all implemented in the pc.world.oldmodel WorldState class. It is the one the vision system updates, so that is why I made the changes there.

The next point in time can be accessed with the method predictNextState, native to the WorldState class